refactor: Move helper to util folder
