+++ tools/serverpod_cli/lib/src/util/pubspec_helpers.dart
